# apps/server/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashyq Bilim API",
    "description": "University LMS / MOOC platform API (v2, Rust rewrite).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v2/activities/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Full activity including content/details/settings.",
        "operationId": "get_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetail"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete an activity; chapter siblings renumber to stay contiguous.",
        "operationId": "delete_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Partial update: name, publish state, content/details/settings, or the\ntype pair (both `activity_type` and `activity_sub_type` together).",
        "operationId": "update_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetail"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/activities/{id}/blocks": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Blocks attached to an activity.",
        "operationId": "list_blocks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Block"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Attach a file block to an activity by claiming a finalized upload whose\npurpose matches the block type (`image`→`block-image`, etc.).",
        "operationId": "create_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "403": {
            "description": "No write access / not your upload",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Upload not finalized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Bad block type or wrong upload purpose",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/activities/{id}/move": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Move an activity within its chapter, or to another chapter of the same\ncourse via `chapter_id`.",
        "operationId": "move_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Moved"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Cross-course move",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/google": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Start Google sign-in: 303 to Google's consent screen.",
        "operationId": "google_start",
        "parameters": [
          {
            "name": "callback",
            "in": "query",
            "description": "Relative return path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to Google"
          }
        }
      }
    },
    "/api/v2/auth/google/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Google redirects here; on success the session cookie is set and the\nbrowser continues to the original callback path.",
        "operationId": "google_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "Opaque state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "Google-side error",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect into the app (or to login with ?error=)"
          }
        }
      }
    },
...
    "/api/v2/auth/mfa/totp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Start TOTP enrollment (secrets are returned exactly once).",
        "operationId": "totp_enroll",
        "responses": {
          "200": {
            "description": "Enrollment secrets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpEnrollment"
                }
              }
            }
          },
          "409": {
            "description": "Already enrolled",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Remove the TOTP authenticator.",
        "operationId": "totp_remove",
        "responses": {
          "204": {
            "description": "Removed (idempotent)"
          }
        }
      }
    },
    "/api/v2/auth/mfa/totp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Activate TOTP with the first code from the authenticator app.",
        "operationId": "totp_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "TOTP active"
          },
          "400": {
            "description": "Invalid code",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The caller's current session (also the cheapest \"am I logged in?\" probe).",
        "operationId": "current_session",
        "responses": {
          "200": {
            "description": "Current session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "No live session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/sessions": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "All live sessions of the caller (handles are non-bearer identifiers).",
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "Live sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No live session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/sessions/{handle}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke one of the caller's own sessions by handle.",
        "operationId": "revoke_session",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "description": "Session handle from the listing",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "404": {
            "description": "No such session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/blocks/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "One block (visibility follows the course).",
        "operationId": "get_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BlockId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Block",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete a block and release its stored file for reaping.",
        "operationId": "delete_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BlockId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}": {
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete a chapter and its activities; siblings renumber to stay contiguous.",
        "operationId": "delete_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Rename/redescribe a chapter.",
        "operationId": "update_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}/activities": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Append an activity to a chapter (type/subtype pair must be valid).",
        "operationId": "create_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created (appended last)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid type/subtype pair",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}/move": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Move a chapter to a new position (clamped; siblings renumber).",
        "operationId": "move_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Moved"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Newest-first collection listing: public plus the caller's own.",
        "operationId": "list_collections",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "next_cursor from the previous page",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CollectionId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 20)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create a collection (requires `collection:create:platform`); every\nattached course must be readable by the caller.",
        "operationId": "create_collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/courses/{id}/chapters": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Append a chapter to a course (course write access).",
        "operationId": "create_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created (appended last)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/rbac/roles": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "All roles with their grants (requires `role:read:platform`).",
        "operationId": "list_roles",
        "responses": {
          "200": {
            "description": "Roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Activity"
          },
          {
            "type": "object",
            "required": [
              "content",
              "details",
              "settings"
            ],
            "properties": {
              "content": {
                "description": "Editor content (dynamic pages) or type-specific payload."
              },
              "details": {},
              "settings": {}
            }
          }
        ],
        "description": "Full activity view with the heavy jsonb columns."
      },
      "ActivityId": {
        "type": "string",
        "format": "uuid"
      },
...
      "TotpVerifyRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateActivityRequest": {
        "type": "object",
        "properties": {
          "activity_sub_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "activity_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Change together with `activity_sub_type` (both or neither)."
          },
          "content": {},
          "details": {},
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "settings": {}
        },
        "additionalProperties": false
      },
...
    }
  },
...
}